Fix tests and remove old code
Removed builders and entriesToDays converter. They were used in the
OvertimeFactory and are no longer needed.

Removed the filter from week.getMinutesWorked because the new
CalendarDataGenerator drops all days that are not in the current month.
Previously the filter was used to dismiss entries not in the current
month. Now the filter parameter makes no sense.